requirements.txt
Pillow
instagrapi
python-dotenv
schedule
